Create README.md
Even though LocoNet has been around for about 25 years, I still would consider it to be the best Layout Control Bus on the market, as to me it represents the best compromise between communication speed and reliability, simplicity of required hardware and related cost, message concept and structuring, and the open and expandable peer-to-peer approach. One small element though that I think is rather weak is the LocoNet FastClock. Luckily the LocoNet Gateway we built in previous videos offers a simple bridge between LocoNet and powerful cloud computing. In this video we try to move the LocoNet FastClock to the cloud and explore what we can use it for on our layout.